💄Add ScalableButtonStyle
