+++ WeHere/Views&ViewModels/Map/MapView.swift
@@ -147,11 +147,3 @@ extension MapView {
         }
     }
 }
-
-struct ScalableButtonStyle: ButtonStyle {
-    func makeBody(configuration: Self.Configuration) -> some View {
-        configuration.label
-            .scaleEffect(configuration.isPressed ? 1.1 : 1)
-            .background(.clear)
-    }
-}
